# openapi.yaml
openapi: 3.0.0
info:
  title: Pet Adoption API
  version: 1.0.0
  description: API for pet adoption
tags:
  - name: Pets
    description: Pet management endpoints
  - name: Shelters
    description: Shelter management endpoints
  - name: Users
    description: User management endpoints
components:
  schemas:
    Pet:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        age:
          type: number
        size:
          type: string
        description:
          type: string
        status:
          type: string
        shelterId:
          type: string
    Shelter:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        preferences:
          type: object
          properties:
            species:
              type: array
              items:
                type: string
            size:
              type: array
              items:
                type: string
paths:
  /pets:
    get:
      tags:
        - Pets
      summary: Get all pets
      parameters:
        - in: query
          name: species
          required: false
          schema:
            type: string
          description: Filter pets by species
      responses:
        '200':
          description: List of all pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
    post:
      tags:
        - Pets
      summary: Create a new pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '201':
          description: Pet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Invalid request body
  /pets/{id}:
    get:
      tags:
        - Pets
      summary: Get a pet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pet details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
    put:
      tags:
        - Pets
      summary: Update a pet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
      responses:
        '200':
          description: Pet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '404':
          description: Pet not found
    delete:
      tags:
        - Pets
      summary: Delete a pet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Pet deleted successfully
        '404':
          description: Pet not found
  /shelters:
    get:
      tags:
        - Shelters
      summary: Get all shelters
      responses:
        '200':
          description: List of all shelters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shelter'
    post:
      tags:
        - Shelters
      summary: Create a new shelter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shelter'
      responses:
        '201':
          description: Shelter created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelter'
  /shelters/{id}:
    get:
      tags:
        - Shelters
      summary: Get a shelter by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Shelter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelter'
    put:
      tags:
        - Shelters
      summary: Update a shelter by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shelter'
      responses:
        '200':
          description: Shelter updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shelter'
    delete:
      tags:
        - Shelters
      summary: Delete a shelter by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Shelter deleted successfully
        '404':
          description: Shelter not found
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - Users
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
        - Users
      summary: Update a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - Users
      summary: Delete a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found